moved TaskToFutureAdapter to util folder and changed some method access controls
