+++ app/src/main/java/ch/epfl/favo/util/TaskToFutureAdapter.java
@@ -1,4 +1,4 @@
-package ch.epfl.favo.common;
+package ch.epfl.favo.util;
 
 import android.annotation.SuppressLint;
 
@@ -7,8 +7,8 @@ import com.google.android.gms.tasks.Task;
 import java.util.concurrent.CompletableFuture;
 
 @SuppressLint("NewApi")
-class TaskToFutureAdapter<T> extends CompletableFuture<T> {
-  TaskToFutureAdapter(Task<T> task) {
+public class TaskToFutureAdapter<T> extends CompletableFuture<T> {
+  public TaskToFutureAdapter(Task<T> task) {
     super();
     task.addOnSuccessListener(this::complete);
     task.addOnFailureListener(this::completeExceptionally);
